ai-ml-dev/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Networking & Protocols
paramiko==3.4.0
scapy==2.5.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.12.0
ruff==0.1.8
flake8==6.1.0

# Utilities
requests==2.31.0
aiofiles==23.2.1
